Modification to return counterexample on UNSAT
In particular, the TLLHypercubeReach object now has a remote method
getCounterExamplePoint that returns a counterexample if one was found.
This should return None if there was no counterexample, and so it should
be safe to call in all circumstances. (Although clearly it was intended
only to be called on UNSAT cases.)