Correctly trap case when quadrant is a singleton point
I hadn't thought of this, but it is bound to happen as the quadrants get
smaller and smaller. Eventually, the quadrant will be so small that it
captures a linear segment of one of the boundary hyperplanes. In this
case the *subsequent* quadrant refinement will have its midpoint exactly
on this line segment -- hence, some of the quadrants will have empty
interiors.