Change default mpi arguments
The latest version of the Mellanox OFED drivers appears to create uverb devices in /dev/infiniband even if the physical hardware isn't present. Thus, these mpi arguments are necessary when using this docker container on a host with the Mellanox OFED drivers installed. This is because mpirun will detect the devices in /dev/infiniband and try to use UCX/RDMA, causing a segfault.

This allows a single host image build on Azure with the OFED drivers installed. There shouldn't be any problems on host systems where OFED is not installed/or where infiniband hardware is actually present