First attempt at allowing depth-wise recursion in parallel
The idea is to alter the recursion so that it pushes towards the corners
first, and instead of returning the box bounding constraints to update a
"global" variable tabulating the current best bounding box so far. Then
by using the charm scheduler, I am allowing different recursive paths to
execute in turn, instead of waiting for one recursive path to "bottom
out". This is intended to ensure that the corners are visited as early
as possible in the hopes that they are providing the most relevant
extents to the bounding box. The hope then, is that the global bounding
box gets close to the final bounding box early, so many box refinements
can be skipped.