Fix an ugly, ugly concurrency bug
For some unknown reason, charm.wait was not, in fact, waiting for all of the futures to become ready. As a result there were remote executions that were continuing past the time the outermost computeLTIBbox call completed, doing who knows what all the while.

I only discovered this because I was trying to build a progress meter, and I couldn't for the life of me get it to report progress for all recursive calls. (See subsequent commit.)

However, this turned out to be **extremely** fortuitous. This concurrency bug was masking a very serious bug in my code regarding the selection of the tolerance for FastBATLLNN. (See subsequent commit.)