Fix significant bug in LTIReach code: tol for FastBATLLNN was too large
This was basically causing infinite recursion, because even as the input "quadrant" to the TLL was getting smaller and smaller, the fixed-size (w.r.t epsilon and B) tolerance for FastBATLLNN caused the returned output bounds to be of fixed width. Thus, it was never possible to refine the bounds on the output of the TLL (by refining the state set) enough to get within the proscribed reachability tolerance.

I believe the calculation implemented here is adequate, but I will revisit it again to be sure.

This bug was masked by the concurrency bug in the commit before last. Basically, for some examples, the infinite recursion was being hidden by the concurrency bug. That is, in some instances the outer function would return, even though there were remote methods doing infinite recursion for the reason stated above. This also finally explains why my initial trials of the lti_reach experiment were so inconsistent: some networks would finish extremely quickly and some would never make any progress. It seems the difference was whether or not the fixed tolerances created this infinite recursion problem, and whether or not the concurrency bug masked it.