Fix a concurrency bug when no refinements are needed in the first call
Another way to say this is that the first call of computeLTIBbox should
never suspend to wait for all of the corners to be refined. This
prevents *itself* from setting cornerRefined and allDone for those
non-descended quadrants. This happens when one of the very first
quadrants does not require descent, and so that main remote method was
deadlocking while waiting for all(self.cornerRefined).