Update poset checking group to match new DistributedHash/Poset behavior
In particular, the incorporated rev. bump on HyperplaneRegionEnum uses a new method to track when all node checks/updates have been completed in poset-checking mode. Thus, the bookkeeping associated with schedCount in this Chare is obviated. Likewise the `getSchedCount` method is no longer needed.

I have also included -- but commented out -- a new init method to give the Node instance access to the poset PE list, and hence allow for `schedRandomPosetPe` call to be disentangled from the check method if desired (there seems to performance benefit to this as of now, though).