Adapt FastBATLLNN to new vectorSet-based flipped constraints object
In particular, deserialize these objects on initialization of the poset
checker group.

This improves performance because it triggers Numba JIT compilation of
vectorSet methods in *parallel* on all poset-checker PEs.

I noticed this problem because I saw significantly degraded performance
for the basic example on very large Azure clusters (e.g. >100 CPUs).
Basically what was happening was that each poset calculation was
effectively sending its nodes to as-yet-unused PEs, which effectively
triggered *serial* compilation of the Numba methods on different PEs.